+++ app/components/History/PotentialRewardsRevealAll/PotentialRewardsRevealAll.tsx
@@ -1,9 +1,9 @@
 "use client";
 
 import { revealQuestion, revealQuestions } from "@/app/actions/chompResult";
-import { RevealProps } from "@/app/hooks/useReveal";
 import { useRevealedContext } from "@/app/providers/RevealProvider";
 import { numberToCurrencyFormatter } from "@/app/utils/currency";
+import { RevealProps } from "@/types/reveal";
 import { useQueryClient } from "@tanstack/react-query";
 import { startTransition, useCallback, useOptimistic, useState } from "react";
